+++ CHANGELOG.md
@@ -17,6 +17,7 @@ OpenNMT-tf follows [semantic versioning 2.0.0](https://semver.org/). The API cov
 ### Fixes and improvements
 
 * Fix error in `SelfAttentionDecoder` when `memory` is not defined (e.g. in LM tasks)
+* Fix `UnicodeDecodeError` when printing predictions on the standard output in Docker containers
 
 ## [1.10.1](https://github.com/OpenNMT/OpenNMT-tf/releases/tag/v1.10.1) (2018-10-15)